+++ src/show.ml
@@ -70,7 +70,7 @@ let trigger_kind (kind : trigger_kind) =
   | INFO -> "INFO"
   | FOLLOWUP -> "FOLLOWUP"
   | SHADOW -> "SHADOW"
-  | WARP target -> fmt "WARP to %s at (%0.2f, %0.2f)" target.room_name target.pos.x target.pos.y
+  | WARP warp -> fmt "WARP to %s at (%0.2f, %0.2f)" warp.room_name warp.target.x warp.target.y
   | CUTSCENE -> "CUTSCENE"
   | RESPAWN -> "RESPAWN"
   | PURPLE_PEN -> "PURPLE_PEN"
@@ -480,3 +480,41 @@ let text_config (config : Interaction.text_config) : string =
   fmt "padding : %s;\nmargin_x : %f;\nmargin_y_top : %f;\nmargin_y_bottom : %f;\ncentered : %b;\n"
     (vector config.padding) config.margin_x config.margin_y_top config.margin_y_bottom
     config.centered
+
+let interaction_step (step : Interaction.step) : string =
+  let open Interaction in
+  match step with
+  | STEP (INITIALIZE_INTERACTIONS _) -> "(INITIALIZE_INTERACTIONS _)"
+  | STEP FADE_SCREEN_OUT -> "FADE_SCREEN_OUT"
+  | STEP FADE_SCREEN_IN -> "FADE_SCREEN_IN"
+  | STEP (SHAKE_SCREEN _) -> "(SHAKE_SCREEN _)"
+  | STEP DEBUG -> "DEBUG"
+  | STEP (WAIT _) -> "(WAIT _)"
+  | STEP (WARP _) -> "(WARP _)"
+  | STEP (DOOR_WARP _) -> "(DOOR_WARP _)"
+  | STEP (SPAWN_VENGEFUL_SPIRIT (_, _, _)) -> "(SPAWN_VENGEFUL_SPIRIT (_, _, _))"
+  | STEP (TEXT _) -> "(TEXT _)"
+  | STEP (FLOATING_TEXT (_, _)) -> "(FLOATING_TEXT (_, _))"
+  | STEP (FOCUS_ABILITY_TEXT (_, _, _)) -> "(FOCUS_ABILITY_TEXT (_, _, _))"
+  | STEP (ABILITY_TEXT (_, _)) -> "(ABILITY_TEXT (_, _))"
+  | STEP RESET_TEXT -> "RESET_TEXT"
+  | STEP (DIALOGUE (_, _)) -> "(DIALOGUE (_, _))"
+  | STEP (PURPLE_PEN_TEXT _) -> "(PURPLE_PEN_TEXT _)"
+  | STEP (SET_FIXED_CAMERA (_, _)) -> "(SET_FIXED_CAMERA (_, _))"
+  | STEP SET_GHOST_CAMERA -> "SET_GHOST_CAMERA"
+  | STEP (SET_CAMERA_MOTION _) -> "(SET_CAMERA_MOTION _)"
+  | STEP (HIDE_LAYER _) -> "(HIDE_LAYER _)"
+  | STEP (UNHIDE_LAYER _) -> "(UNHIDE_LAYER _)"
+  | STEP HIDE_BOSS_DOORS -> "HIDE_BOSS_DOORS"
+  | STEP UNHIDE_BOSS_DOORS -> "UNHIDE_BOSS_DOORS ->"
+  | CURRENT_GHOST FILL_LIFE_VAPOR -> "GHOST FILL_LIFE_VAPOR"
+  | CURRENT_GHOST (INCREASE_HEALTH_TEXT _) -> "GHOST (INCREASE_HEALTH_TEXT _)"
+  | CURRENT_GHOST (SET_POSE _) -> "GHOST (SET_POSE _)"
+  | CURRENT_GHOST (ADD_ITEM _) -> "GHOST (ADD_ITEM _)"
+  | CURRENT_GHOST UNSET_FLOOR -> "GHOST UNSET_FLOOR"
+  | CURRENT_GHOST (ENTITY _) -> "GHOST (ENTITY _)"
+  | CURRENT_GHOST (PARTY _) -> "GHOST (PARTY _) ->"
+  | PARTY_GHOST (_, _)
+  | ENEMY (_, _)
+  | NPC (_, _) ->
+    ""